feat: add hooks for querys
